// frontend/src/app/api/model/role.ts
/**
 * Planning Poker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const Role = {
    SCRUM_MASTER: 'SCRUM_MASTER',
    DEVELOPER: 'DEVELOPER'
} as const;
export type Role = typeof Role[keyof typeof Role];