// frontend/src/app/api/model/session.ts
/**
 * Planning Poker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Vote } from './vote';
import { Participant } from './participant';
import { SessionStatus } from './sessionStatus';
import { Round } from './round';


export interface Session { 
    id?: string;
    scrumMasterId?: string;
    gitlabProjectId?: string;
    participants?: Array<Participant>;
    completedRounds?: Array<Round>;
    votes?: { [key: string]: Vote; };
    currentIssueIid?: number;
    status?: SessionStatus;
}
export namespace Session {
}